requirements.txt
aiogram>=3.13,<4
python-dotenv>=1.0
Pillow>=10.0
